requirements.txt
amqp==2.6.1
asgiref==3.5.2
asn1crypto==1.0.1
async-timeout==4.0.2
Babel==2.7.0
bcrypt==3.1.7
billiard==3.6.4.0
celery==4.3.0
certifi==2019.9.11
cffi==1.15.0
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
cryptography==2.7
defusedxml==0.6.0
Deprecated==1.2.13
diff-match-patch==20181111
Django==2.2.8
django-background-tasks==1.2.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-compat==1.0.15
django-guardian==2.1.0
django-import-export==1.2.0
django-request-logging==0.7.0
django-rest-swagger==2.2.0
django-timezone-field==3.1
djangorestframework==3.10.3
et-xmlfile==1.0.1
Fabric3==1.14.post1
factory-boy==2.12.0
Faker==2.0.4
flower==0.9.3
idna==2.8
importlib-metadata==0.23
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
kombu==4.6.11
MarkupPy==1.14
MarkupSafe==1.1.1
more-itertools==7.2.0
msgpack==1.0.4
munch==2.5.0
odfpy==1.4.0
openapi-codec==1.3.2
openpyxl==3.0.2
packaging==21.3
parameterized==0.7.0
paramiko==2.6.0
pillow==9.2.0
postgres==4.0
prompt-toolkit==3.0.30
psycopg2-pool==1.1
pycparser==2.19
PyNaCl==1.3.0
pyparsing==3.0.9
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2022.1
PyYAML==5.1.2
redis==3.3.11
requests==2.22.0
sentry-sdk==0.12.3
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
tablib==0.14.0
text-unidecode==1.3
tornado==5.1.1
uritemplate==3.0.0
urllib3==1.25.6
uWSGI==2.0.20
vine==1.3.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0
